+++ pennylane/transforms/control.py
@@ -15,6 +15,8 @@
 Contains the control transform.
 """
 from functools import wraps
+
+import pennylane as qml
 from pennylane.tape import QuantumTape, get_active_tape
 from pennylane.operation import Operation, AnyWires
 from pennylane.wires import Wires
@@ -45,9 +47,24 @@ def expand_with_control(tape, control_wire):
                 # pylint: disable=protected-access
                 op._controlled(control_wire)
             else:
-                tmp_tape = op.expand()
+                # Attempt to decompose the operation and apply
+                # controls to each gate in the decomposition.
+                with new_tape.stop_recording():
+                    try:
+                        tmp_tape = op.expand()
+
+                    except NotImplementedError:
+                        # No decomposition is defined. Create a
+                        # ControlledQubitUnitary gate using the operation
+                        # matrix representation.
+                        with QuantumTape() as tmp_tape:
+                            qml.ControlledQubitUnitary(
+                                op.matrix, control_wires=control_wire, wires=op.wires
+                            )
+
                 tmp_tape = expand_with_control(tmp_tape, control_wire)
                 requeue_ops_in_tape(tmp_tape)
+
     return new_tape