+++ tests/transforms/test_control.py
@@ -210,3 +210,48 @@ def test_no_control_defined():
     # Check that all operations are updated to their controlled version.
     for op in tape.operations:
         assert type(op) in {qml.ControlledPhaseShift, qml.Toffoli, qml.CRX, qml.CSWAP}
+
+
+def test_no_decomposition_defined():
+    """Test that a controlled gate that has no control transform defined,
+    as well as no decomposition transformed defined, still works correctly"""
+
+    with QuantumTape() as tape:
+        ctrl(qml.CZ, 0)(wires=[1, 2])
+
+    tape = expand_tape(tape)
+
+    assert len(tape.operations) == 1
+    assert tape.operations[0].name == "ControlledQubitUnitary"
+
+
+def test_controlled_template():
+    """Test that a controlled template correctly expands
+    on a device that doesn't support it"""
+
+    weights = np.ones([3, 2])
+
+    with QuantumTape() as tape:
+        ctrl(qml.templates.BasicEntanglerLayers, 0)(weights, wires=[1, 2])
+
+    tape = expand_tape(tape)
+    assert len(tape.operations) == 9
+    assert all(o.name in {"CRX", "Toffoli"} for o in tape.operations)
+
+
+def test_controlled_template_and_operations():
+    """Test that a combination of controlled templates and operations correctly expands
+    on a device that doesn't support it"""
+
+    weights = np.ones([3, 2])
+
+    def ansatz(weights, wires):
+        qml.PauliX(wires=wires[0])
+        qml.templates.BasicEntanglerLayers(weights, wires=wires)
+
+    with QuantumTape() as tape:
+        ctrl(ansatz, 0)(weights, wires=[1, 2])
+
+    tape = expand_tape(tape)
+    assert len(tape.operations) == 10
+    assert all(o.name in {"CNOT", "CRX", "Toffoli"} for o in tape.operations)