+++ .github/CHANGELOG.md
@@ -351,6 +351,10 @@ random_mat2 = rng.standard_normal(3, requires_grad=False)
 
 <h3>Bug fixes</h3>
 
+* Fixes a bug where `qml.ctrl` would fail to transform gates that had no
+  control defined and no decomposition defined.
+  [(#1376)](https://github.com/PennyLaneAI/pennylane/pull/1376)
+
 * Copying the `JacobianTape` now correctly also copies the `jacobian_options` attribute. This fixes a bug
   allowing the JAX interface to support adjoint differentiation.
   [(#1349)](https://github.com/PennyLaneAI/pennylane/pull/1349)